AbstractEthereumService and new EthereumService singleton
