+++ src/entity/EventSubscription.ts
@@ -21,7 +21,7 @@ import { Logger } from '../Logger';
 import { MetricsReporter } from '../MetricsReporter';
 import { AsyncPoll } from '../AsyncPoll';
 import { arrayChunker } from '../utils';
-import { EthereumService } from '../eth/EthereumService';
+import { AbstractEthereumService } from '../eth/AbstractEthereumService';
 
 const request = require('request');
 const requestPromise = require('request-promise-native');
@@ -344,7 +344,7 @@ export class EventSubscription extends BaseEntity {
             const fromBlock: number = eventSubscription.lastBlock ? +eventSubscription.lastBlock + 1 : 0;
 
             const startTime = Date.now();
-            const ethereumService: EthereumService = await eventSubscription.contractEntity.network.getEthereumService();
+            const ethereumService: AbstractEthereumService = await eventSubscription.contractEntity.network.getEthereumService();
 
             try {
                 const logs = await ethereumService.getContractEvents(
